data-pipeline/requirements.txt
requests>=2.31
psycopg2-binary>=2.9
python-dotenv>=1.0
google-api-python-client>=2.100
tenacity>=8.2
